Rename pallene.lexer to pallene.Lexer
