+++ pallene-dev-1.rockspec
@@ -29,7 +29,7 @@ build = {
       ["pallene.coder"] =          "pallene/coder.lua",
       ["pallene.gc"] =             "pallene/gc.lua",
       ["pallene.ir"] =             "pallene/ir.lua",
-      ["pallene.lexer"] =          "pallene/lexer.lua",
+      ["pallene.Lexer"] =          "pallene/Lexer.lua",
       ["pallene.Location"] =       "pallene/Location.lua",
       ["pallene.parser"] =         "pallene/parser.lua",
       ["pallene.print_ir"] =       "pallene/print_ir.lua",
